scrape/requirements.txt
requests==2.32.3
firecrawl-py==1.9.0
beautifulsoup4==4.12.3
google-generativeai==0.8.3
